Split viewable and reactive modules
