+++ panel/viewable.py
@@ -1,14 +1,15 @@
 """
-Defines the Viewable and Reactive baseclasses allow all panel objects
-to display themselves, communicate with a Python process and react in
-response to changes to parameters and the underlying bokeh models.
-"""
-from __future__ import absolute_import, division, unicode_literals
+Defines the baseclasses that make a component render to a bokeh model
+and become viewable including:
 
-import difflib
+* Layoutable: Defines parameters concerned with layout and style
+* ServableMixin: Mixin class that defines methods to serve object on server
+* Renderable: Defines methods to render a component as a bokeh model
+* Viewable: Defines methods to view the component in the
+  notebook, on the server or in static exports
+"""
 import logging
 import sys
-import threading
 import traceback
 import uuid
 
@@ -20,25 +21,27 @@ import param
 from bokeh.document.document import Document as _Document
 from bokeh.io import curdoc as _curdoc
 from pyviz_comms import JupyterCommManager
-from tornado import gen
 
-from .callbacks import PeriodicCallback
 from .config import config, panel_extension
 from .io.embed import embed_state
-from .io.model import add_to_doc, hold, patch_cds_msg
+from .io.model import add_to_doc, patch_cds_msg
 from .io.notebook import (
-    ipywidget, push, render_mimebundle, render_model, show_embed, show_server
+    ipywidget, render_mimebundle, render_model, show_embed, show_server
 )
 from .io.save import save
 from .io.state import state
-from .io.server import StoppableThread, get_server, unlocked
-from .util import edit_readonly, escape, param_reprs
+from .io.server import StoppableThread, get_server
+from .util import escape, param_reprs
 
 
 LinkWatcher = namedtuple("Watcher","inst cls fn mode onlychanged parameter_names what queued target links transformed")
 
 
 class Layoutable(param.Parameterized):
+    """
+    Layoutable defines shared style and layout related parameters
+    for all Panel components with a visual representation.
+    """
 
     align = param.ObjectSelector(default='start',
                                  objects=['start', 'end', 'center'], doc="""
@@ -201,6 +204,8 @@ class Layoutable(param.Parameterized):
             provided aspect ratio.
     """)
 
+    __abstract = True
+
     def __init__(self, **params):
         if (params.get('width', None) is not None and
             params.get('height', None) is not None and
@@ -215,6 +220,9 @@ class Layoutable(param.Parameterized):
 
 
 class ServableMixin(object):
+    """
+    Mixin to define methods shared by objects which can served.
+    """
 
     def _modify_doc(self, server_id, title, doc, location):
         """
@@ -346,6 +354,14 @@ class ServableMixin(object):
 
 
 class Renderable(param.Parameterized):
+    """
+    Baseclass for objects which can be rendered to a Bokeh model.
+
+    It therefore declare APIs for initializing the models from
+    parameter values.
+    """
+
+    __abstract = True
 
     def __init__(self, **params):
         super(Renderable, self).__init__(**params)
@@ -413,6 +429,10 @@ class Renderable(param.Parameterized):
             add_to_doc(model, doc)
         return model
 
+    def _init_properties(self):
+        return {k: v for k, v in self.param.get_param_values()
+                if v is not None}
+
     def get_root(self, doc=None, comm=None):
         """
         Returns the root model and applies pre-processing hooks
@@ -435,39 +455,15 @@ class Renderable(param.Parameterized):
         state._views[ref] = (self, root, doc, comm)
         return root
 
-    #----------------------------------------------------------------
-    # Public API
-    #----------------------------------------------------------------
-
-    def clone(self, **params):
-        """
-        Makes a copy of the object sharing the same parameters.
-
-        Arguments
-        ---------
-        params: Keyword arguments override the parameters on the clone.
-
-        Returns
-        -------
-        Cloned Viewable object
-        """
-        return type(self)(**dict(self.param.get_param_values(), **params))
-
-    def pprint(self):
-        """
-        Prints a compositional repr of the class.
-        """
-        print(self)
-
 
 class Viewable(Renderable, Layoutable, ServableMixin):
     """
-    Viewable is the baseclass all objects in the panel library are
-    built on. It defines the interface for declaring any object that
-    displays itself by transforming the object(s) being wrapped into
-    models that can be served using bokeh's layout engine. The class
-    also defines various methods that allow Viewable objects to be
-    displayed in the notebook and on bokeh server.
+    Viewable is the baseclass all visual components in the panel
+    library are built on. It defines the interface for declaring any
+    object that displays itself by transforming the object(s) being
+    wrapped into models that can be served using bokeh's layout
+    engine. The class also defines various methods that allow Viewable
+    objects to be displayed in the notebook and on bokeh server.
     """
 
     _preprocessing_hooks = []
@@ -537,49 +533,6 @@ class Viewable(Renderable, Layoutable, ServableMixin):
             return render_model(model)
         return render_mimebundle(model, doc, comm, manager)
 
-    def _comm_change(self, doc, ref, attr, old, new):
-        if attr in self._changing.get(ref, []):
-            self._changing[ref].remove(attr)
-            return
-
-        with hold(doc):
-            self._process_events({attr: new})
-
-    def _server_change(self, doc, ref, attr, old, new):
-        if attr in self._changing.get(ref, []):
-            self._changing[ref].remove(attr)
-            return
-
-        state._locks.clear()
-        self._events.update({attr: new})
-        if not self._processing:
-            self._processing = True
-            if doc.session_context:
-                doc.add_timeout_callback(partial(self._change_coroutine, doc), self._debounce)
-            else:
-                self._change_event(doc)
-
-    def _process_events(self, events):
-        self.param.set_param(**self._process_property_change(events))
-
-    @gen.coroutine
-    def _change_coroutine(self, doc=None):
-        self._change_event(doc)
-
-    def _change_event(self, doc=None):
-        try:
-            state.curdoc = doc
-            thread = threading.current_thread()
-            thread_id = thread.ident if thread else None
-            state._thread_id = thread_id
-            events = self._events
-            self._events = {}
-            self._process_events(events)
-        finally:
-            self._processing = False
-            state.curdoc = None
-            state._thread_id = None
-
     def _server_destroy(self, session_context):
         """
         Server lifecycle hook triggered when session is destroyed.
@@ -591,6 +544,26 @@ class Viewable(Renderable, Layoutable, ServableMixin):
     #----------------------------------------------------------------
     # Public API
     #----------------------------------------------------------------
+
+    def clone(self, **params):
+        """
+        Makes a copy of the object sharing the same parameters.
+
+        Arguments
+        ---------
+        params: Keyword arguments override the parameters on the clone.
+
+        Returns
+        -------
+        Cloned Viewable object
+        """
+        return type(self)(**dict(self.param.get_param_values(), **params))
+
+    def pprint(self):
+        """
+        Prints a compositional repr of the class.
+        """
+        print(self)
 
     def select(self, selector=None):
         """
@@ -731,422 +704,3 @@ class Viewable(Renderable, Layoutable, ServableMixin):
             loc_model = loc._get_model(doc, model)
             doc.add_root(loc_model)
         return doc
-
-
-class Syncable(Renderable):
-
-    # Timeout if a notebook comm message is swallowed
-    _timeout = 20000
-
-    # Timeout before the first event is processed
-    _debounce = 50
-
-    # Mapping from parameter name to bokeh model property name
-    _rename = {}
-
-    __abstract = True
-
-    events = []
-
-    def __init__(self, **params):
-        super(Syncable, self).__init__(**params)
-        self._processing = False
-        self._events = {}
-        self._callbacks = []
-        self._links = []
-        self._link_params()
-        self._changing = {}
-
-    # Allows defining a mapping from model property name to a JS code
-    # snippet that transforms the object before serialization
-    _js_transforms = {}
-
-    # Transforms from input value to bokeh property value
-    _source_transforms = {}
-    _target_transforms = {}
-
-    #----------------------------------------------------------------
-    # Callback API
-    #----------------------------------------------------------------
-
-    def _update_model(self, events, msg, root, model, doc, comm):
-        self._changing[root.ref['id']] = [
-            attr for attr, value in msg.items()
-            if not model.lookup(attr).property.matches(getattr(model, attr), value)
-        ]
-        try:
-            model.update(**msg)
-        finally:
-            del self._changing[root.ref['id']]
-
-    def param_change(self, *events):
-        msgs = []
-        for event in events:
-            msg = self._process_param_change({event.name: event.new})
-            if msg:
-                msgs.append(msg)
-
-        events = {event.name: event for event in events}
-        msg = {k: v for msg in msgs for k, v in msg.items()}
-        if not msg:
-            return
-
-        for ref, (model, parent) in self._models.items():
-            if ref not in state._views or ref in state._fake_roots:
-                continue
-            viewable, root, doc, comm = state._views[ref]
-            if comm or not doc.session_context or state._unblocked(doc):
-                with unlocked():
-                    self._update_model(events, msg, root, model, doc, comm)
-                if comm and 'embedded' not in root.tags:
-                    push(doc, comm)
-            else:
-                cb = partial(self._update_model, events, msg, root, model, doc, comm)
-                doc.add_next_tick_callback(cb)
-
-    def _link_params(self):
-        params = self._synced_params()
-        if params:
-            watcher = self.param.watch(self.param_change, params)
-            self._callbacks.append(watcher)
-
-    def _link_props(self, model, properties, doc, root, comm=None):
-        ref = root.ref['id']
-        if config.embed:
-            return
-
-        for p in properties:
-            if isinstance(p, tuple):
-                _, p = p
-            if comm:
-                model.on_change(p, partial(self._comm_change, doc, ref))
-            else:
-                model.on_change(p, partial(self._server_change, doc, ref))
-
-    def _process_events(self, events):
-        with edit_readonly(self):
-            self.param.set_param(**self._process_property_change(events))
-
-    @property
-    def _linkable_params(self):
-        return [p for p in self._synced_params()
-                if self._source_transforms.get(p, False) is not None]
-
-    def _synced_params(self):
-        return list(self.param)
-
-    def _process_property_change(self, msg):
-        """
-        Transform bokeh model property changes into parameter updates.
-        Should be overridden to provide appropriate mapping between
-        parameter value and bokeh model change. By default uses the
-        _rename class level attribute to map between parameter and
-        property names.
-        """
-        inverted = {v: k for k, v in self._rename.items()}
-        return {inverted.get(k, k): v for k, v in msg.items()}
-
-    def _process_param_change(self, msg):
-        """
-        Transform parameter changes into bokeh model property updates.
-        Should be overridden to provide appropriate mapping between
-        parameter value and bokeh model change. By default uses the
-        _rename class level attribute to map between parameter and
-        property names.
-        """
-        properties = {self._rename.get(k, k): v for k, v in msg.items()
-                      if self._rename.get(k, False) is not None}
-        if 'width' in properties and self.sizing_mode is None:
-            properties['min_width'] = properties['width']
-        if 'height' in properties and self.sizing_mode is None:
-            properties['min_height'] = properties['height']
-        return properties
-
-    #----------------------------------------------------------------
-    # Model API
-    #----------------------------------------------------------------
-
-    def _init_properties(self):
-        return {k: v for k, v in self.param.get_param_values()
-                if v is not None}
-
-
-    def _cleanup(self, root):
-        super(Syncable, self)._cleanup(root)
-        ref = root.ref['id']
-        self._models.pop(ref, None)
-        comm, client_comm = self._comms.pop(ref, (None, None))
-        if comm:
-            try:
-                comm.close()
-            except Exception:
-                pass
-        if client_comm:
-            try:
-                client_comm.close()
-            except Exception:
-                pass
-
-    #----------------------------------------------------------------
-    # Public API
-    #----------------------------------------------------------------
-
-    def link(self, target, callbacks=None, **links):
-        """
-        Links the parameters on this object to attributes on another
-        object in Python. Supports two modes, either specify a mapping
-        between the source and target object parameters as keywords or
-        provide a dictionary of callbacks which maps from the source
-        parameter to a callback which is triggered when the parameter
-        changes.
-
-        Arguments
-        ---------
-        target: object
-          The target object of the link.
-        callbacks: dict
-          Maps from a parameter in the source object to a callback.
-        **links: dict
-          Maps between parameters on this object to the parameters
-          on the supplied object.
-        """
-        if links and callbacks:
-            raise ValueError('Either supply a set of parameters to '
-                             'link as keywords or a set of callbacks, '
-                             'not both.')
-        elif not links and not callbacks:
-            raise ValueError('Declare parameters to link or a set of '
-                             'callbacks, neither was defined.')
-
-        _updating = []
-        def link(*events):
-            for event in events:
-                if event.name in _updating: continue
-                _updating.append(event.name)
-                try:
-                    if callbacks:
-                        callbacks[event.name](target, event)
-                    else:
-                        setattr(target, links[event.name], event.new)
-                except Exception:
-                    raise
-                finally:
-                    _updating.pop(_updating.index(event.name))
-        params = list(callbacks) if callbacks else list(links)
-        cb = self.param.watch(link, params)
-        link = LinkWatcher(*tuple(cb)+(target, links, callbacks is not None))
-        self._links.append(link)
-        return cb
-
-    def controls(self, parameters=[], jslink=True):
-        """
-        Creates a set of widgets which allow manipulating the parameters
-        on this instance. By default all parameters which support
-        linking are exposed, but an explicit list of parameters can
-        be provided.
-
-        Arguments
-        ---------
-        parameters: list(str)
-           An explicit list of parameters to return controls for.
-        jslink: bool
-           Whether to use jslinks instead of Python based links.
-           This does not allow using all types of parameters.
-
-        Returns
-        -------
-        A layout of the controls
-        """
-        from .param import Param
-        from .layout import Tabs, WidgetBox
-        from .widgets import LiteralInput
-
-        if parameters:
-            linkable = parameters
-        elif jslink:
-            linkable = self._linkable_params
-        else:
-            linkable = list(self.param)
-
-        params = [p for p in linkable if p not in Layoutable.param]
-        controls = Param(self.param, parameters=params, default_layout=WidgetBox,
-                         name='Controls')
-        layout_params = [p for p in linkable if p in Layoutable.param]
-        if 'name' not in layout_params and self._rename.get('name', False) is not None and not parameters:
-            layout_params.insert(0, 'name')
-        style = Param(self.param, parameters=layout_params, default_layout=WidgetBox,
-                      name='Layout')
-        if jslink:
-            for p in params:
-                widget = controls._widgets[p]
-                widget.jslink(self, value=p, bidirectional=True)
-                if isinstance(widget, LiteralInput):
-                    widget.serializer = 'json'
-            for p in layout_params:
-                widget = style._widgets[p]
-                widget.jslink(self, value=p, bidirectional=True)
-                if isinstance(widget, LiteralInput):
-                    widget.serializer = 'json'
-
-        if params and layout_params:
-            return Tabs(controls.layout[0], style.layout[0])
-        elif params:
-            return controls.layout[0]
-        return style.layout[0]
-
-    def jscallback(self, args={}, **callbacks):
-        """
-        Allows defining a JS callback to be triggered when a property
-        changes on the source object. The keyword arguments define the
-        properties that trigger a callback and the JS code that gets
-        executed.
-
-        Arguments
-        ----------
-        args: dict
-          A mapping of objects to make available to the JS callback
-        **callbacks: dict
-          A mapping between properties on the source model and the code
-          to execute when that property changes
-
-        Returns
-        -------
-        callback: Callback
-          The Callback which can be used to disable the callback.
-        """
-
-        from .links import Callback
-        for k, v in list(callbacks.items()):
-            callbacks[k] = self._rename.get(v, v)
-        return Callback(self, code=callbacks, args=args)
-
-    def jslink(self, target, code=None, args=None, bidirectional=False, **links):
-        """
-        Links properties on the source object to those on the target
-        object in JS code. Supports two modes, either specify a
-        mapping between the source and target model properties as
-        keywords or provide a dictionary of JS code snippets which
-        maps from the source parameter to a JS code snippet which is
-        executed when the property changes.
-
-        Arguments
-        ----------
-        target: HoloViews object or bokeh Model or panel Viewable
-          The target to link the value to.
-        code: dict
-          Custom code which will be executed when the widget value
-          changes.
-        bidirectional: boolean
-          Whether to link source and target bi-directionally
-        **links: dict
-          A mapping between properties on the source model and the
-          target model property to link it to.
-
-        Returns
-        -------
-        link: GenericLink
-          The GenericLink which can be used unlink the widget and
-          the target model.
-        """
-        if links and code:
-            raise ValueError('Either supply a set of properties to '
-                             'link as keywords or a set of JS code '
-                             'callbacks, not both.')
-        elif not links and not code:
-            raise ValueError('Declare parameters to link or a set of '
-                             'callbacks, neither was defined.')
-        if args is None:
-            args = {}
-
-        mapping = code or links
-        for k in mapping:
-            if k.startswith('event:'):
-                continue
-            elif k not in self.param and k not in list(self._rename.values()):
-                matches = difflib.get_close_matches(k, list(self.param))
-                if matches:
-                    matches = ' Similar parameters include: %r' % matches
-                else:
-                    matches = ''
-                raise ValueError("Could not jslink %r parameter (or property) "
-                                 "on %s object because it was not found.%s"
-                                 % (k, type(self).__name__, matches))
-            elif (self._source_transforms.get(k, False) is None or
-                  self._rename.get(k, False) is None):
-                raise ValueError("Cannot jslink %r parameter on %s object, "
-                                 "the parameter requires a live Python kernel "
-                                 "to have an effect." % (k, type(self).__name__))
-
-        if isinstance(target, Syncable) and code is None:
-            for k, p in mapping.items():
-                if k.startswith('event:'):
-                    continue
-                elif p not in target.param and p not in list(target._rename.values()):
-                    matches = difflib.get_close_matches(p, list(target.param))
-                    if matches:
-                        matches = ' Similar parameters include: %r' % matches
-                    else:
-                        matches = ''
-                    raise ValueError("Could not jslink %r parameter (or property) "
-                                     "on %s object because it was not found.%s"
-                                    % (p, type(self).__name__, matches))
-                elif (target._source_transforms.get(p, False) is None or
-                      target._rename.get(p, False) is None):
-                    raise ValueError("Cannot jslink %r parameter on %s object "
-                                     "to %r parameter on %s object. It requires "
-                                     "a live Python kernel to have an effect."
-                                     % (k, type(self).__name__, p, type(target).__name__))
-
-        from .links import Link
-        return Link(self, target, properties=links, code=code, args=args,
-                    bidirectional=bidirectional)
-
-
-class Reactive(Syncable, Viewable):
-    """
-    Reactive is a Viewable object that also supports syncing between
-    the objects parameters and the underlying bokeh model either via
-    the defined pyviz_comms.Comm type or when using bokeh server.
-
-    In order to bi-directionally link parameters with bokeh model
-    instances the _link_params and _link_props methods define
-    callbacks triggered when either the parameter or bokeh property
-    values change. Since there may not be a 1-to-1 mapping between
-    parameter and the model property the _process_property_change and
-    _process_param_change may be overridden to apply any necessary
-    transformations.
-    """
-
-    #----------------------------------------------------------------
-    # Public API
-    #----------------------------------------------------------------
-
-    def add_periodic_callback(self, callback, period=500, count=None,
-                              timeout=None, start=True):
-        """
-        Schedules a periodic callback to be run at an interval set by
-        the period. Returns a PeriodicCallback object with the option
-        to stop and start the callback.
-
-        Arguments
-        ---------
-        callback: callable
-          Callable function to be executed at periodic interval.
-        period: int
-          Interval in milliseconds at which callback will be executed.
-        count: int
-          Maximum number of times callback will be invoked.
-        timeout: int
-          Timeout in seconds when the callback should be stopped.
-        start: boolean (default=True)
-          Whether to start callback immediately.
-
-        Returns
-        -------
-        Return a PeriodicCallback object with start and stop methods.
-        """
-        cb = PeriodicCallback(callback=callback, period=period,
-                              count=count, timeout=timeout)
-        if start:
-            cb.start()
-        return cb
